feat(payment): PAYMENTS-11577 make field rendering a reusable component for future use
